Add SEO, OpenGraph, structured data, and LLM discovery files
- Expand Next.js metadata with keywords, OpenGraph, Twitter cards, and robots directives
- Add JSON-LD structured data (SoftwareSourceCode schema)
- Refactor page.tsx to server component for metadata export support
- Add robots.ts, sitemap.ts, and dynamic OG image generation
- Add favicon using Lucide text-cursor-input icon
- Create llms.txt and llms-full.txt for LLM/AI search discoverability
- Target keywords: textarea with tags, prompt textarea, shadcn textarea, AI chat input
