+++ app/robots.ts
@@ -1,0 +1,13 @@
+import type { MetadataRoute } from 'next'
+
+export default function robots(): MetadataRoute.Robots {
+  return {
+    rules: [
+      {
+        userAgent: '*',
+        allow: '/',
+      },
+    ],
+    sitemap: 'https://prompt-area.com/sitemap.xml',
+  }
+}
